almost done updating
