+++ modules/net.plantabyte.drptrace/src/net/plantabyte/drptrace/IntervalTracer.java
@@ -31,49 +31,45 @@ import static net.plantabyte.drptrace.trace.TraceMachine.followEdge;
  * curve nodes, with a higher number resulting in fewer nodes. 10 is usually a
  * good value to use.
  */
-public class IntervalTracer {
+public class IntervalTracer extends Tracer {
+
+	private int interval;
+
+	/**
+	 * Constructs a new <code>IntervalTracer</code> with the given precision interval. The
+	 * 	interval number is the ratio of provided path points to the number of
+	 * 	beziers. For example, a interval of 10 means that there will be 1 bezier
+	 * 	for every 10 path points.
+	 * @param interval Controls the density of beziers (higher number means
+	 * 	                  fewer bezier curves). MUST be at least 1 (10 recommended for
+	 * 	                  relatively small or detailed rasters, 50+ for large rasters).
+	 * @throws IllegalArgumentException Thrown if any of the input arguments are
+	 * invalid
+	 */
+	public IntervalTracer(int interval){
+		if(interval <= 1) throw new IllegalArgumentException(String.format("Invalid interval score: %s (must be greator than 1)", interval));
+		this.interval = interval;
+	}
 	/**
 	 * Traces a series of points as a sequence of bezier curves, looping back to
-	 * the beginning to form a closed loop. The density of bezier curves is
-	 * controlled by the <code>interval</code> score. Specifically, the
+	 * the beginning to form a closed loop if <code>closedLoop</code> is true.
+	 * The density of bezier curves is controlled by the <code>interval</code>
+	 * score that was set in the constructor. Specifically, the
 	 * interval number is the ratio of provided path points to the number of
 	 * beziers. For example, a interval of 10 means that there will be 1 bezier
 	 * for every 10 path points.
 	 * @param pathPoints A series of points to trace with bezier curves. MUST
-	 *                   contain at least 3 points
-	 * @param interval Controls the density of beziers (higher number means
-	 *                   fewer bezier curves). MUST be at least 1 (10 recommended).
+	 *                   contain at least 3 points for a closed loop or 2 points
+	 *                   for an open trace
+	 * @param closedLoop If true, traceback to the starting point (index 0);
+	 *                   if false, trace to the final point
 	 * @return Returns a list of <code>BezierCurve</code>s tracing the path of
 	 * the points
 	 * @throws IllegalArgumentException Thrown if any of the input arguments are
-	 * invalid
+	 * invalid (eg too few points)
 	 */
-	public BezierShape traceClosedPath(Vec2[] pathPoints, int interval)
-			throws IllegalArgumentException {
-		return tracePath(pathPoints, interval, true);
-	}
-
-	/**
-	 * Traces a series of points as a sequence of bezier curves. The density of
-	 * bezier curves is controlled by the <code>interval</code> score.
-	 * Specifically, the interval number is the ratio of provided path points
-	 * beziers. For example, a interval of 10 means that there will be 1 bezier
-	 * for every 10 path points.
-	 * @param pathPoints A series of points to trace with bezier curves. MUST
-	 *                   contain at least 2 points
-	 * @param interval Controls the density of beziers (higher number means
-	 *                   fewer bezier curves). MUST be at least 1 (10 recommended).
-	 * @return Returns a list of <code>BezierCurve</code>s tracing the path of
-	 * the points
-	 * @throws IllegalArgumentException Thrown if any of the input arguments are
-	 * invalid
-	 */
-	public BezierShape traceOpenPath(Vec2[] pathPoints, int interval)
-			throws IllegalArgumentException {
-		return tracePath(pathPoints, interval, false);
-	}
-
-	private BezierShape tracePath(Vec2[] pathPoints, int interval, boolean closedLoop)
+	@Override
+	public BezierShape tracePath(Vec2[] pathPoints, boolean closedLoop)
 			throws IllegalArgumentException{
 		//
 		final int min_beziers = closedLoop ? 2 : 1;
@@ -113,47 +109,5 @@ public class IntervalTracer {
 		return beziers;
 	}
 
-	/**
-	 * Traces every shape (including the background) of the provided raster bitmap.
-	 * The density of bezier curves is controlled by the <code>interval</code>
-	 * score. Specifically, the interval number is the ratio of provided path
-	 * points beziers. For example, a interval of 10 means that there will be 1
-	 * bezier for every 10 path points.
-	 * @param bitmap A 2D array of integer values, such that each contiguous area
-	 *               of a number is considered to be a single shape.
-	 * @param interval Controls the density of beziers (higher number means
-	 *                   fewer bezier curves). MUST be at least 1 (10 recommended).
-	 * @return Returns a list of <code>BezierShape</code> objects, each representing
-	 * one shape from the raster. The order is important: the shapes should be drawn
-	 * in the order such that the first index is in the back and the last index is
-	 * in the front.
-	 * @throws IllegalArgumentException Thrown if any of the input arguments are
-	 * invalid
-	 */
-	public List<BezierShape> traceAllShapes(final IntMap bitmap, final int interval) throws IllegalArgumentException {
-		final int w = bitmap.getWidth(), h = bitmap.getHeight();
-		var searchedMap = new ZOrderBinaryMap(w, h);
-		var output = new LinkedList<BezierShape>();
-		// algorithm: flood fill each patch, and for
-		// each flood fill, trace the outer edge
-		for(int y = 0; y < h; y++){
-			for(int x = 0; x < w; x++){
-				if(searchedMap.get(x, y) == 0){ // pixel not yet searched
-					// first, find a patch of color in the bitmap
-					int color = bitmap.get(x, y);
-					// next, trace the outer perimeter of the color patch
-					var circumference = followEdge(bitmap, x, y);
-					var vectorized = traceClosedPath(circumference, interval);
-					vectorized.setColor(color);
-					vectorized.setClosed(true);
-					output.add(vectorized);
-					// finally, flood-fill the patch in the searched map
-					floodFill(bitmap, searchedMap, x, y);
-					searchedMap.set(x, y, (byte)1);
-				}
-			}
-		}
-		return output;
-	}
 
 }
